+++ matrix.hpp
@@ -30,14 +30,14 @@
 namespace filter::matrix {
     // function prototypes
     void f3x3matrixAeqI(float A[][3]);
-    void fmatrixAeqI(float* A[], int16 rc);
+    void fmatrixAeqI(float* A[], int rc);
     void f3x3matrixAeqScalar(float A[][3], float Scalar);
     void f3x3matrixAeqInvSymB(float A[][3], float B[][3]);
     void f3x3matrixAeqAxScalar(float A[][3], float Scalar);
     void f3x3matrixAeqMinusA(float A[][3]);
     float f3x3matrixDetA(float A[][3]);
-    void eigencompute(float A[][10], float eigval[], float eigvec[][10], int8 n);
-    void fmatrixAeqInvA(float* A[], int8 iColInd[], int8 iRowInd[], int8 iPivot[], int8 isize);
+    void eigencompute(float A[][10], float eigval[], float eigvec[][10], int n);
+    void fmatrixAeqInvA(float* A[], int iColInd[], int iRowInd[], int iPivot[], int isize);
     void fmatrixAeqRenormRotA(float A[][3]);
 }  // namespace filter::matrix
 #endif  // #ifndef MATRIX_HPP
